Add SpvIcon component & refactor RescanButton & RescanButtonCancel
