+++ app/components/SideBar/MenuBottom/SpvIcon/SpvIcon.module.css
@@ -1,0 +1,9 @@
+.spvIcon {
+  background-image: var(--menu-spvon-icon);
+  height: 2.2rem;
+  width: 2.2rem; 
+}
+
+.spvIcon.off {
+  background-image: var(--menu-spvoff-icon);
+}
