+++ app/components/buttons/AutoBuyerSwitch.jsx
