+++ app/components/buttons/RescanButton/RescanCancelButton.jsx
@@ -1,16 +1,15 @@
 import { FormattedMessage as T } from "react-intl";
 import { Tooltip } from "shared";
-
-const rescanBtnMessage = "Cancel rescan";
+import styles from "./RescanButton.module.css";
 
 export default ({ rescanRequest, rescanCancel }) => (
   <Tooltip
-    text={<T id="sidebar.rescanCancelBtn.tip" m={rescanBtnMessage} />}
+    text={<T id="sidebar.rescanCancelBtn.tip" m="Cancel rescan" />}
     disabled={!rescanRequest}>
     <button
       aria-label="Cancel rescan"
       disabled={!rescanRequest}
-      className={"rescan-cancel-button"}
+      className={styles.cancel}
       onClick={() => rescanCancel()}
     />
   </Tooltip>
