+++ app/style/themes/lightTheme.js
@@ -141,7 +141,8 @@ const lightTheme = {
   "menu-hamburger": "url('style/icons/hamburger.svg')",
   "menu-arrow": "url('style/icons/arrow.svg')",
   "menu-rescan-icon": "url('style/icons/menu-rescan.svg')",
-  "menu-peers-icon": "url('style/icons/peers.svg')",
+  "menu-spvon-icon": "url('style/icons/menu-spvon.svg')",
+  "menu-spvoff-icon": "url('style/icons/menu-spvoff.svg')",
   "tab-icon-balance": "url('style/icons/balanceSelector.svg')",
   "tab-icon-transactions": "url('style/icons/transactions-default.png')",
   "tab-icon-tickets": "url('style/icons/tickets-default.png')",
